requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.7.1
bs4==0.0.1
CacheControl==0.12.5
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
discord.py==1.0.0
firebase-admin==2.16.0
google-api-core==1.8.2
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
grpcio==1.19.0
idna==2.8
idna-ssl==1.1.0
lxml==4.3.3
msgpack==0.6.1
multidict==4.5.2
Pillow==5.4.1
protobuf==3.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.9
typing-extensions==3.7.2
urllib3==1.24.1
websockets==6.0
wikipedia==1.4.0
yarl==1.3.0
